tests/requirements.txt
pytest==8.3.2
pyyaml==6.0.2

